+++ frontend/src/pages/inbounds/form/InboundFormModal.tsx
@@ -22,7 +22,6 @@ import {
   ArrowDownOutlined,
   ArrowUpOutlined,
   DeleteOutlined,
-  MinusOutlined,
   PlusOutlined,
 } from '@ant-design/icons';
 
@@ -47,18 +46,10 @@ import {
 } from '@/schemas/forms/inbound-form';
 import { antdRule } from '@/utils/zodForm';
 import {
-  ALPN_OPTION,
-  Address_Port_Strategy,
-  DOMAIN_STRATEGY_OPTION,
   Protocols,
   SNIFFING_OPTION,
-  TCP_CONGESTION_OPTION,
-  UTLS_FINGERPRINT,
 } from '@/schemas/primitives';
-import {
-  HappyEyeballsSchema,
-  SockoptStreamSettingsSchema,
-} from '@/schemas/protocols/stream/sockopt';
+import { SockoptStreamSettingsSchema } from '@/schemas/protocols/stream/sockopt';
 import { HysteriaStreamSettingsSchema } from '@/schemas/protocols/stream/hysteria';
 import { TlsStreamSettingsSchema } from '@/schemas/protocols/security/tls';
 import { RealityStreamSettingsSchema } from '@/schemas/protocols/security/reality';
@@ -86,10 +77,12 @@ import {
   WireguardFields,
 } from './protocols';
 import {
+  ExternalProxyForm,
   GrpcForm,
   HttpUpgradeForm,
   KcpForm,
   RawForm,
+  SockoptForm,
   WsForm,
   XhttpForm,
 } from './transport';
@@ -1058,378 +1051,9 @@ export default function InboundFormModal({
 
       {network === 'kcp' && <KcpForm />}
 
-      <Form.Item
-        noStyle
-        shouldUpdate={(prev, curr) => {
-          const a = (prev.streamSettings as { externalProxy?: unknown[] } | undefined)?.externalProxy;
-          const b = (curr.streamSettings as { externalProxy?: unknown[] } | undefined)?.externalProxy;
-          return (Array.isArray(a) ? a.length : 0) !== (Array.isArray(b) ? b.length : 0);
-        }}
-      >
-        {({ getFieldValue }) => {
-          const arr = getFieldValue(['streamSettings', 'externalProxy']);
-          const on = Array.isArray(arr) && arr.length > 0;
-          return (
-            <>
-              <Form.Item label={t('pages.inbounds.form.externalProxy')}>
-                <Switch checked={on} onChange={toggleExternalProxy} />
-              </Form.Item>
-              {on && (
-                <Form.List name={['streamSettings', 'externalProxy']}>
-                  {(fields, { add, remove }) => (
-                    <>
-                      <Form.Item label=" " colon={false}>
-                        <Button
-                          size="small"
-                          type="primary"
-                          onClick={() => add({
-                            forceTls: 'same',
-                            dest: '',
-                            port: 443,
-                            remark: '',
-                            sni: '',
-                            fingerprint: '',
-                            alpn: [],
-                          })}
-                        >
-                          <PlusOutlined />
-                        </Button>
-                      </Form.Item>
-                      <Form.Item wrapperCol={{ span: 24 }}>
-                        {fields.map((field) => (
-                          <div key={field.key} style={{ margin: '8px 0' }}>
-                            <Space.Compact block>
-                              <Form.Item name={[field.name, 'forceTls']} noStyle>
-                                <Select
-                                  style={{ width: '20%' }}
-                                  options={[
-                                    { value: 'same', label: t('pages.inbounds.same') },
-                                    { value: 'none', label: t('none') },
-                                    { value: 'tls', label: 'TLS' },
-                                  ]}
-                                />
-                              </Form.Item>
-                              <Form.Item name={[field.name, 'dest']} noStyle>
-                                <Input style={{ width: '30%' }} placeholder={t('host')} />
-                              </Form.Item>
-                              <Form.Item name={[field.name, 'port']} noStyle>
-                                <InputNumber style={{ width: '15%' }} min={1} max={65535} />
-                              </Form.Item>
-                              <Form.Item name={[field.name, 'remark']} noStyle>
-                                <Input style={{ width: '25%' }} placeholder={t('pages.inbounds.remark')} />
-                              </Form.Item>
-                              <InputAddon onClick={() => remove(field.name)}>
-                                <MinusOutlined />
-                              </InputAddon>
-                            </Space.Compact>
-                            <Form.Item
-                              noStyle
-                              shouldUpdate={(prev, curr) =>
-                                prev.streamSettings?.externalProxy?.[field.name]?.forceTls
-                                !== curr.streamSettings?.externalProxy?.[field.name]?.forceTls
-                              }
-                            >
-                              {({ getFieldValue }) => {
-                                const ft = getFieldValue([
-                                  'streamSettings', 'externalProxy', field.name, 'forceTls',
-                                ]);
-                                if (ft !== 'tls') return null;
-                                return (
-                                  <Space.Compact style={{ marginTop: 6 }} block>
-                                    <Form.Item name={[field.name, 'sni']} noStyle>
-                                      <Input style={{ width: '30%' }} placeholder={t('pages.inbounds.form.sniPlaceholder')} />
-                                    </Form.Item>
-                                    <Form.Item name={[field.name, 'fingerprint']} noStyle>
-                                      <Select
-                                        style={{ width: '30%' }}
-                                        placeholder={t('pages.inbounds.form.fingerprint')}
-                                        options={[
-                                          { value: '', label: t('pages.inbounds.form.defaultOption') },
-                                          ...Object.values(UTLS_FINGERPRINT).map((fp) => ({
-                                            value: fp,
-                                            label: fp,
-                                          })),
-                                        ]}
-                                      />
-                                    </Form.Item>
-                                    <Form.Item name={[field.name, 'alpn']} noStyle>
-                                      <Select
-                                        mode="multiple"
-                                        style={{ width: '40%' }}
-                                        placeholder="ALPN"
-                                        options={Object.values(ALPN_OPTION).map((a) => ({
-                                          value: a,
-                                          label: a,
-                                        }))}
-                                      />
-                                    </Form.Item>
-                                  </Space.Compact>
-                                );
-                              }}
-                            </Form.Item>
-                          </div>
-                        ))}
-                      </Form.Item>
-                    </>
-                  )}
-                </Form.List>
-              )}
-            </>
-          );
-        }}
-      </Form.Item>
+      <ExternalProxyForm toggleExternalProxy={toggleExternalProxy} />
 
-      <Form.Item
-        noStyle
-        shouldUpdate={(prev, curr) => {
-          const a = (prev.streamSettings as { sockopt?: object } | undefined)?.sockopt;
-          const b = (curr.streamSettings as { sockopt?: object } | undefined)?.sockopt;
-          return !!a !== !!b;
-        }}
-      >
-        {({ getFieldValue }) => {
-          const sock = getFieldValue(['streamSettings', 'sockopt']);
-          const on = !!sock && typeof sock === 'object' && Object.keys(sock).length > 0;
-          return (
-            <>
-              <Form.Item label="Sockopt">
-                <Switch checked={on} onChange={toggleSockopt} />
-              </Form.Item>
-              {on && (
-                <>
-                  <Form.Item name={['streamSettings', 'sockopt', 'mark']} label={t('pages.inbounds.form.routeMark')}>
-                    <InputNumber min={0} />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpKeepAliveInterval']}
-                    label={t('pages.inbounds.form.tcpKeepAliveInterval')}
-                  >
-                    <InputNumber min={0} />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpKeepAliveIdle']}
-                    label={t('pages.inbounds.form.tcpKeepAliveIdle')}
-                  >
-                    <InputNumber min={0} />
-                  </Form.Item>
-                  <Form.Item name={['streamSettings', 'sockopt', 'tcpMaxSeg']} label={t('pages.inbounds.form.tcpMaxSeg')}>
-                    <InputNumber min={0} />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpUserTimeout']}
-                    label={t('pages.inbounds.form.tcpUserTimeout')}
-                  >
-                    <InputNumber min={0} />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpWindowClamp']}
-                    label={t('pages.inbounds.form.tcpWindowClamp')}
-                  >
-                    <InputNumber min={0} />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'acceptProxyProtocol']}
-                    label={t('pages.inbounds.form.proxyProtocol')}
-                    valuePropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpFastOpen']}
-                    label={t('pages.inbounds.form.tcpFastOpen')}
-                    valuePropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpMptcp']}
-                    label={t('pages.inbounds.form.multipathTcp')}
-                    valuePropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'penetrate']}
-                    label={t('pages.inbounds.form.penetrate')}
-                    valuePropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'V6Only']}
-                    label={t('pages.inbounds.form.v6Only')}
-                    valuePropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'domainStrategy']}
-                    label={t('pages.xray.wireguard.domainStrategy')}
-                  >
-                    <Select
-                      style={{ width: '50%' }}
-                      options={Object.values(DOMAIN_STRATEGY_OPTION).map((d) => ({ value: d, label: d }))}
-                    />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'tcpcongestion']}
-                    label={t('pages.inbounds.form.tcpCongestion')}
-                  >
-                    <Select
-                      style={{ width: '50%' }}
-                      options={Object.values(TCP_CONGESTION_OPTION).map((c) => ({ value: c, label: c }))}
-                    />
-                  </Form.Item>
-                  <Form.Item name={['streamSettings', 'sockopt', 'tproxy']} label="TProxy">
-                    <Select
-                      style={{ width: '50%' }}
-                      options={[
-                        { value: 'off', label: 'Off' },
-                        { value: 'redirect', label: 'Redirect' },
-                        { value: 'tproxy', label: 'TProxy' },
-                      ]}
-                    />
-                  </Form.Item>
-                  <Form.Item name={['streamSettings', 'sockopt', 'dialerProxy']} label={t('pages.inbounds.form.dialerProxy')}>
-                    <Input />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'interfaceName']}
-                    label={t('pages.inbounds.info.interfaceName')}
-                  >
-                    <Input />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'trustedXForwardedFor']}
-                    label={t('pages.inbounds.form.trustedXForwardedFor')}
-                  >
-                    <Select
-                      mode="tags"
-                      style={{ width: '100%' }}
-                      tokenSeparators={[',']}
-                      options={[
-                        { value: 'CF-Connecting-IP', label: 'CF-Connecting-IP' },
-                        { value: 'X-Real-IP', label: 'X-Real-IP' },
-                        { value: 'True-Client-IP', label: 'True-Client-IP' },
-                        { value: 'X-Client-IP', label: 'X-Client-IP' },
-                      ]}
-                    />
-                  </Form.Item>
-                  <Form.Item
-                    name={['streamSettings', 'sockopt', 'addressPortStrategy']}
-                    label={t('pages.inbounds.form.addressPortStrategy')}
-                  >
-                    <Select
-                      style={{ width: '50%' }}
-                      options={Object.values(Address_Port_Strategy).map((v) => ({ value: v, label: v }))}
-                    />
-                  </Form.Item>
-                  <Form.Item shouldUpdate noStyle>
-                    {({ getFieldValue, setFieldValue }) => {
-                      const he = getFieldValue(['streamSettings', 'sockopt', 'happyEyeballs']);
-                      const hasHe = he != null;
-                      return (
-                        <>
-                          <Form.Item label="Happy Eyeballs">
-                            <Switch
-                              checked={hasHe}
-                              onChange={(v) => {
-                                setFieldValue(
-                                  ['streamSettings', 'sockopt', 'happyEyeballs'],
-                                  v ? HappyEyeballsSchema.parse({}) : undefined,
-                                );
-                              }}
-                            />
-                          </Form.Item>
-                          {hasHe && (
-                            <>
-                              <Form.Item
-                                name={['streamSettings', 'sockopt', 'happyEyeballs', 'tryDelayMs']}
-                                label={t('pages.inbounds.form.tryDelayMs')}
-                              >
-                                <InputNumber min={0} placeholder="0 disabled — 250 recommended" />
-                              </Form.Item>
-                              <Form.Item
-                                name={['streamSettings', 'sockopt', 'happyEyeballs', 'prioritizeIPv6']}
-                                label={t('pages.inbounds.form.prioritizeIPv6')}
-                                valuePropName="checked"
-                              >
-                                <Switch />
-                              </Form.Item>
-                              <Form.Item
-                                name={['streamSettings', 'sockopt', 'happyEyeballs', 'interleave']}
-                                label={t('pages.inbounds.form.interleave')}
-                              >
-                                <InputNumber min={1} />
-                              </Form.Item>
-                              <Form.Item
-                                name={['streamSettings', 'sockopt', 'happyEyeballs', 'maxConcurrentTry']}
-                                label={t('pages.inbounds.form.maxConcurrentTry')}
-                              >
-                                <InputNumber min={0} />
-                              </Form.Item>
-                            </>
-                          )}
-                        </>
-                      );
-                    }}
-                  </Form.Item>
-                  <Form.List name={['streamSettings', 'sockopt', 'customSockopt']}>
-                    {(fields, { add, remove }) => (
-                      <>
-                        <Form.Item label={t('pages.inbounds.form.customSockopt')}>
-                          <Button
-                            type="dashed"
-                            size="small"
-                            onClick={() => add({ type: 'int', level: '6', opt: '', value: '' })}
-                          >
-                            + {t('pages.inbounds.form.addCustomOption')}
-                          </Button>
-                        </Form.Item>
-                        {fields.map((field) => (
-                          <Space.Compact key={field.key} style={{ display: 'flex', marginBottom: 8 }}>
-                            <Form.Item name={[field.name, 'system']} noStyle>
-                              <Select
-                                placeholder="all"
-                                allowClear
-                                style={{ width: 100 }}
-                                options={[
-                                  { value: 'linux', label: 'linux' },
-                                  { value: 'windows', label: 'windows' },
-                                  { value: 'darwin', label: 'darwin' },
-                                ]}
-                              />
-                            </Form.Item>
-                            <Form.Item name={[field.name, 'type']} noStyle>
-                              <Select
-                                style={{ width: 80 }}
-                                options={[
-                                  { value: 'int', label: 'int' },
-                                  { value: 'str', label: 'str' },
-                                ]}
-                              />
-                            </Form.Item>
-                            <Form.Item name={[field.name, 'level']} noStyle>
-                              <Input placeholder="level (6=TCP)" style={{ width: 100 }} />
-                            </Form.Item>
-                            <Form.Item name={[field.name, 'opt']} noStyle>
-                              <Input placeholder="opt" style={{ width: 120 }} />
-                            </Form.Item>
-                            <Form.Item name={[field.name, 'value']} noStyle>
-                              <Input placeholder="value" style={{ flex: 1 }} />
-                            </Form.Item>
-                            <Button danger onClick={() => remove(field.name)}>−</Button>
-                          </Space.Compact>
-                        ))}
-                      </>
-                    )}
-                  </Form.List>
-                </>
-              )}
-            </>
-          );
-        }}
-      </Form.Item>
+      <SockoptForm toggleSockopt={toggleSockopt} />
 
       <FinalMaskForm
         name={['streamSettings', 'finalmask']}
