+++ frontend/src/pages/inbounds/form/transport/index.ts
@@ -4,3 +4,5 @@ export { default as GrpcForm } from './grpc';
 export { default as XhttpForm } from './xhttp';
 export { default as HttpUpgradeForm } from './httpupgrade';
 export { default as KcpForm } from './kcp';
+export { default as ExternalProxyForm } from './external-proxy';
+export { default as SockoptForm } from './sockopt';
