+++ frontend/src/test/inbound-form-blocks.test.tsx
@@ -3,10 +3,12 @@ import { Form, type FormInstance } from 'antd';
 import type { ReactNode } from 'react';
 
 import {
+  ExternalProxyForm,
   GrpcForm,
   HttpUpgradeForm,
   KcpForm,
   RawForm,
+  SockoptForm,
   WsForm,
   XhttpForm,
 } from '@/pages/inbounds/form/transport';
@@ -14,13 +16,22 @@ import { RealityForm, TlsForm } from '@/pages/inbounds/form/security';
 import type { InboundFormValues } from '@/schemas/forms/inbound-form';
 import { renderWithProviders, fieldLabels } from './test-utils';
 
-function FormHarness({ children }: { children: (form: FormInstance<InboundFormValues>) => ReactNode }) {
+function FormHarness({
+  children,
+  initialValues,
+}: {
+  children: (form: FormInstance<InboundFormValues>) => ReactNode;
+  initialValues?: Record<string, unknown>;
+}) {
   const [form] = Form.useForm<InboundFormValues>();
-  return <Form form={form}>{children(form)}</Form>;
+  return <Form form={form} initialValues={initialValues}>{children(form)}</Form>;
 }
 
-function renderInForm(node: (form: FormInstance<InboundFormValues>) => ReactNode) {
-  return renderWithProviders(<FormHarness>{node}</FormHarness>);
+function renderInForm(
+  node: (form: FormInstance<InboundFormValues>) => ReactNode,
+  initialValues?: Record<string, unknown>,
+) {
+  return renderWithProviders(<FormHarness initialValues={initialValues}>{node}</FormHarness>);
 }
 
 const noop = () => {};
@@ -53,6 +64,34 @@ describe('inbound transport forms', () => {
 
   it('XhttpForm field structure is stable', () => {
     renderInForm((form) => <XhttpForm form={form} />);
+    expect(fieldLabels()).toMatchSnapshot();
+  });
+
+  it('ExternalProxyForm field structure is stable (one TLS entry)', () => {
+    renderInForm(
+      () => <ExternalProxyForm toggleExternalProxy={noop} />,
+      {
+        streamSettings: {
+          externalProxy: [{
+            forceTls: 'tls',
+            dest: '',
+            port: 443,
+            remark: '',
+            sni: '',
+            fingerprint: '',
+            alpn: [],
+          }],
+        },
+      },
+    );
+    expect(fieldLabels()).toMatchSnapshot();
+  });
+
+  it('SockoptForm field structure is stable (enabled + happy eyeballs)', () => {
+    renderInForm(
+      () => <SockoptForm toggleSockopt={noop} />,
+      { streamSettings: { sockopt: { happyEyeballs: {} } } },
+    );
     expect(fieldLabels()).toMatchSnapshot();
   });
 });
